+++ lib/TFVuTrading/PanelOrderButtons_structs.cpp
@@ -38,16 +38,16 @@ double PanelOrderButtons_Order::PositionExitStopValue() const {
   return boost::lexical_cast<double>( m_sPositionExitStopValue );
 }
 
-size_t PanelOrderButtons_Order::QuanStock() const {
-  return boost::lexical_cast<size_t>( m_sQuanStock );
+uint32_t PanelOrderButtons_Order::QuanStock() const {
+  return boost::lexical_cast<uint32_t>( m_sQuanStock );
 }
 
-size_t PanelOrderButtons_Order::QuanFuture() const {
-  return boost::lexical_cast<size_t>( m_sQuanFuture );
+uint32_t PanelOrderButtons_Order::QuanFuture() const {
+  return boost::lexical_cast<uint32_t>( m_sQuanFuture );
 }
 
-size_t PanelOrderButtons_Order::QuanOption() const {
-  return boost::lexical_cast<size_t>( m_sQuanOption );
+uint32_t PanelOrderButtons_Order::QuanOption() const {
+  return boost::lexical_cast<uint32_t>( m_sQuanOption );
 }
 
 } // namespace tf
